Rename module to config
