+++ contracts/finality/src/contract.rs
@@ -1,5 +1,5 @@
 use crate::error::ContractError;
-use crate::exec::admin::set_enabled;
+use crate::exec::config::set_enabled;
 use crate::exec::finality::{handle_finality_signature, handle_public_randomness_commit};
 use crate::msg::BabylonMsg;
 use crate::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
