+++ contracts/finality/src/exec/mod.rs
@@ -1,2 +1,2 @@
-pub mod admin;
+pub mod config;
 pub mod finality;
